+++ ntex/src/util/mod.rs
@@ -1,6 +1,5 @@
 pub mod buffer;
 pub mod counter;
-pub mod either;
 mod extensions;
 pub mod inflight;
 pub mod keepalive;
@@ -9,5 +8,4 @@ pub mod time;
 pub mod timeout;
 pub mod variant;
 
-pub use self::either::either;
 pub use self::extensions::Extensions;