+++ ntex/CHANGES.md
@@ -6,6 +6,8 @@
 
 * server: Rename ServerBulder::system_exit to stop_runtime
 
+* util: Drop Either service, use Variant instead
+
 ## [0.2.0-b.12] - 2021-02-18
 
 * http: Fix KeepAlive::Os support for h1 dispatcher